Move configs folder into src so customer can just grap src folder if trying to use this project quickly
