+++ src/configs/ui.js
